// src/app/enums/deviceManagment.enum.ts
export enum DeviceModes {
    rollback = 'ROLLBACK',
    skuTransfer = 'SKUTRANSFER',
    esnSwap = 'ESNSWAP',
    softwareVersion='SOFTWAREVERSION',
    routeCalculate = 'ROUTERECALCULATE',
    dataElements = 'DATAELEMENTS',
    esnRemove = 'ESNREMOVE',
    esnRevive = 'ESNREVIVE',
    wmxToRmx = 'WMX-RMX',
    wmxToSku = 'WMX-SKU',
    fullCase = 'FULLCASE',
    rmxTransfer = 'RMXTRANSFER',
    wmxTransfer = 'WMXTRANSFER',
    serialNumber = 'SERIALNUMBER',
    process = 'Process',
    processAll = 'Process All',
    single = 'SINGLE',
    disposition = 'Disposition',
    IVCCode = 'IVCCode',
    }