// src/app/enums/storage.enum.ts
export enum StorageData {
    clientData = 'clientData',
    sessionClose = 'sessionClose',
    sessionStorageClose = 'sessionStorageClose',
    module = 'Module',
    testSerialNumber = 'testSerialNumber',
    addWho = 'AddWho',
    token = 'token',
    rolesSiteIds = 'rolesSiteIds',
    currentProfile = 'currentProfile',
    siteIds = 'siteIds',
    clientId = 'clientId',
    location = 'location',
    defaultLang = 'defaultLang',
    logoutData = 'logoutData',
    receiptSearchObj = 'receiptSearchObj',
    receivingObj = 'receivingObj',
    authKey = 'Auth_Key',
    langConfig = 'langConfig',
    timeOutConfig = 'timeOutConfig',
    release = 'release',
    clientImage = 'clientImage',
    imageAltText = 'imageAltText',
    menu = 'menu',
    messages = 'messages',
    controlConfig = 'controlConfig',
    appConfig= 'appConfig',
    rolesList = 'rolesList',
    operation = 'Operation',
    userProfileChanged = 'userProfileChanged',
    timmerCount = 'timmerCount',
    continue = 'continue',
    patternList = 'patternList',
    statusTypesList = 'statusTypesList',
    authObj="authObj",
    env = 'env'
    }